refactor(controller): extract deduplication helpers
